Import changes for goma client
  - effdeb35443feab4d10ae2f8ba728f2970f902ca Fix on "Expose compiler_proxy_status_logo.h"
  - a7dc457101c70d800c8f8a5137b7f2f590e7d869 Expose compiler_proxy_status_logo.h
  - 3c57c284fb0a62dec28139d4713c918958e6cee0 Expose HandleStatusLogo
  - 54c72878c32218cbabda1c481deb60805440dfbf Treat Windows paths as case-insensitive when removing dup...
  - 01473813ec1d2fc551b187fd7ee920d1e0b3e639 Run python 2to3 on goma_ctl.py
  - 47bc619fc0da1a9f98dc5ef8fa5d7b7c53da18eb fix: should wait inflight OAuth2 token refresh request.
  - e75812e2da331e6c51086adcf805d6d41b898ec3 Roll client/third_party/boringssl/src bf5021a6b..9cde848bd
  - 2be956f96a31c137ec672407d60ceb093069c9ba Workaround for presubmit bot failure
  - 65cb0cd3877973301907aa801e1db8d3198cc306 compiler_proxy logo is not internal only
  - f1e7bdab864ee93a8f1a78e6cbbff754e703f82e Fix exec req normalizer for cl.exe
  - 47847592eeac25dc661e9ac32cd8476c4d5ee30e Remove DONT_KILL_COMMANDS
  - f34726bbd58d7ab99adecfbf4ab03896a23b56b6 Collect clang-cl.exe executable binaries

GitOrigin-RevId: effdeb35443feab4d10ae2f8ba728f2970f902ca

+++ test/simpletry.sh
@@ -460,11 +460,20 @@ CURRENT_DIR_BACKUP=$PWD
 cd $LOCAL_CXX_DIR
 assert_success "${LOCAL_CXX} $CURRENT_DIR_BACKUP/test/oneinclude.cc \
   -c -o $CURRENT_DIR_BACKUP/out_plain.o"
-expect_success "gomacc_relative_path_${CXX}" \
+
+
+if [ "$(uname)" = "Darwin" ]; then
+    # TODO: In Mac, failed to include <iostream>. b/122490524
+    MAYBE_FAIL="FAIL_"
+fi
+
+expect_success "${MAYBE_FAIL}gomacc_relative_path_${CXX}" \
      "${GOMACC} ./${CXX} $CURRENT_DIR_BACKUP/test/oneinclude.cc \
      -c -o $CURRENT_DIR_BACKUP/out.o"
-expect_success "${HERMETIC_GCC}gomacc_relative_path_${CXX}_oneinclude.o" \
+expect_success "${MAYBE_FAIL}${HERMETIC_GCC}gomacc_relative_path_${CXX}_oneinclude.o" \
     "objcmp ${CURRENT_DIR_BACKUP}/out_plain.o ${CURRENT_DIR_BACKUP}/out.o"
+MAYBE_FAIL=
+
 cd $CURRENT_DIR_BACKUP
 
 rm -f out2.o out_plain.o
@@ -523,19 +532,6 @@ if (${LOCAL_CXX} -DTHREAD_SANITIZER -fsanitize=thread -fPIC \
    "${GOMA_CXX} -DTHREAD_SANITIZER -fsanitize=thread -fPIC \
     -fsanitize-blacklist=test/tsan-ign.txt \
     -o out.o -c test/oneinclude.cc"
-fi
-
-if [ "$CXX" = "clang++" ]; then
-  # See: b/16826568
-  ext=".so"
-  if [ "$(uname -s)" == "Darwin" ]; then
-    ext=".dylib"
-  fi
-
-  expect_success "${CXX}_load_plugin_in_relative_path" \
-  "${GOMACC} ${LOCAL_CXX} -Xclang -load -Xclang \
-   third_party/llvm-build/Release+Asserts/lib/libFindBadConstructs${ext} \
-   -o out.o -c test/oneinclude.cc"
 fi
 
 if [ "$CXX" = "g++" ]; then
@@ -746,9 +742,23 @@ if [ "$(uname)" = "Linux" ]; then
   fi
 fi
 
-if [ "$GOMATEST_USE_CHROMIUM_CLANG" = "1" ]; then
-  # automatically detects .so and .dylib.
-  CLANG_PLUGIN=$(echo $(dirname $LOCAL_CXX)/../lib/libFindBadConstructs.*)
+
+ext=".so"
+if [ "$(uname -s)" == "Darwin" ]; then
+    ext=".dylib"
+fi
+CLANG_PLUGIN="third_party/llvm-build/Release+Asserts/lib/libFindBadConstructs${ext}"
+
+# TODO: Remove this test? (b/122436038)
+if [ "$CXX" = "clang++" -a -f "${CLANG_PLUGIN}" ]; then
+  # See: b/16826568
+  expect_success "${CXX}_load_plugin_in_relative_path" \
+  "${GOMACC} ${LOCAL_CXX} -Xclang -load -Xclang ${CLANG_PLUGIN} \
+   -o out.o -c test/oneinclude.cc"
+fi
+
+# TODO: Remove plugin test? (b/122436038)
+if [ "$GOMATEST_USE_CHROMIUM_CLANG" = "1" -a -f "${CLANG_PLUGIN}" ]; then
   CLANG_PLUGIN_BASE=$(basename $CLANG_PLUGIN)
 
   cp -p ${CLANG_PLUGIN} ./${CLANG_PLUGIN_BASE}
